fix: CI 编译失败(prompt .txt 被 gitignore 排除) + state 平铺对齐 Python + clippy 零 warning
- .gitignore: *.txt 白名单 rust/crates/rpg-agents/src/prompts/*.txt,
  修复 include_str!() 编译时找不到文件导致 CI 失败
- core.rs /api/state: 把 snapshot.data 平铺到顶层(player/world/memory/...),
  对齐 Python status_payload(),修复前端右栏 player 字段显示"—"
- game.rs build_status_payload: 同样平铺 state_data 而非嵌套在 "state" key 下
- 14 个 clippy warning 全部修复(unused imports / redundant closures /
  unnecessary casts / too_many_arguments / collapsible if-let / doc indent /
  unwrap_or_default / identity map / identical if blocks)

+++ rust/crates/rpg-state/src/structured.rs
@@ -112,7 +112,7 @@ pub fn apply_structured_updates(
             .as_ref()
             .and_then(|p| p.get("turn"))
             .and_then(Value::as_i64)
-            .map(|t| t == current_turn as i64)
+            .map(|t| t == current_turn)
             .unwrap_or(false);
         if player_pending_this_turn {
             asking_for_confirm = true;
